requirements.txt
finmind
pandas
tqdm
tabulate
yfinance
python-dotenv